refactor: cleaned duplicate code
