+++ backend/onyx/chat/process_message.py
@@ -98,14 +98,13 @@ from onyx.file_store.models import InMemoryChatFile
 from onyx.file_store.utils import load_all_chat_files
 from onyx.file_store.utils import save_files
 from onyx.kg.clustering.clustering import kg_clustering
-from onyx.kg.configuration import populate_default_account_employee_definitions
-from onyx.kg.configuration import populate_default_grounded_entity_types
 from onyx.kg.extractions.extraction_processing import kg_extraction
 from onyx.kg.resets.reset_extractions import reset_extraction_kg_index
 from onyx.kg.resets.reset_index import reset_full_kg_index
 from onyx.kg.resets.reset_normalizations import reset_normalization_kg_index
 from onyx.kg.resets.reset_source import reset_source_kg_index
 from onyx.kg.resets.reset_vespa import reset_vespa_kg_index
+from onyx.kg.setup.kg_default_entity_definitions import populate_default_entity_types
 from onyx.llm.exceptions import GenAIDisabledException
 from onyx.llm.factory import get_llms_for_persona
 from onyx.llm.factory import get_main_llm_from_tuple
@@ -618,8 +617,7 @@ def stream_chat_message_objects(
             raise Exception("Vespa KG index reset done")
 
         elif new_msg_req.message == "kg_setup":
-            populate_default_grounded_entity_types()
-            populate_default_account_employee_definitions()
+            populate_default_entity_types()
             raise Exception("KG setup done")
 
     try:
